feat: sync umt_rust with TypeScript implementation (#585)
* feat: sync umt_rust with TypeScript implementation

Add missing Rust functions (getObjectsCommon, getObjectsDiff, gcd_many),
fix precision handling (sum, average, convert_time, addition), correct
deg_to_rad to use f64, and add nCr early return for r==0 || n==r.


* fix: resolve lint errors in getObjectsDiff.ts


* style: apply lint formatting to getObjectsCommon and tests


---------

+++ package/umt_rust/src/math/deg_to_rad.rs
@@ -1,4 +1,4 @@
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 /// Converts degrees to radians.
 ///
@@ -11,6 +11,6 @@ use std::f32::consts::PI;
 /// # Returns
 ///
 /// The angle in radians.
-pub fn umt_deg_to_rad(degrees: f32) -> f32 {
+pub fn umt_deg_to_rad(degrees: f64) -> f64 {
     degrees * PI / 180.0
 }
